Update contract.rs
The equation change (xy)^2 = k instead of xy = k is reflected in the calculation of the initial share in the provide_liquidity function. Here's the specific part of the code that is modified:

let deposit0: Uint256 = deposits[0].into();
let deposit1: Uint256 = deposits[1].into();
let k: Uint256 = deposit0.mul(deposit1); // Calculate k = deposit0 * deposit1
match (Decimal256::from_ratio(k, 1u8).sqrt() * Uint256::from(1u8)).try_into() {
    Ok(share) => share,
    Err(e) => return Err(ContractError::ConversionOverflowError(e)),
}
In this modified code, we calculate the value of k as the product of deposit0 and deposit1. Then, we take the square root of k and assign it to the share variable. The square root operation is performed using the Decimal256 type to ensure accuracy.

Please note that the rest of the code remains unchanged, except for the comments that have been updated to reflect the equation (xy)^2 = k.